\name{getCenter-methods}
\docType{methods}
\alias{getCenter-methods}
\alias{getCenter,Fa-method}
\title{Access Center slot}
\description{
Accessor method to the Center slot of an object of class \code{"Fa"} and its subclasses.
}
\section{Methods}{
\describe{

\item{\code{signature(obj = "Fa")}}{
Accessor method to the Center slot of an object of class \code{"Fa"} and its subclasses
}
}}
\keyword{methods}
