\name{getEigenvalues-methods}
\docType{methods}
\alias{getEigenvalues-methods}
\alias{getEigenvalues,Fa-method}
\title{Access Eigenvalues slot}
\description{
Accessor method to the Eigenvalues slot of an object of class \code{"Fa"} and its subclasses.
}
\section{Methods}{
\describe{

\item{\code{signature(obj = "Fa")}}{
Accessor method to the Eigenvalues slot of an object of class \code{"Fa"} and its subclasses
}
}}
\keyword{methods}
