\name{getFa-methods}
\docType{methods}
\alias{getFa-methods}
\alias{getFa,Fa-method}
\alias{getFa}

\title{Access slots of "Fa"}
\description{
Accessor method to some slots of an object of class \code{"Fa"} and its subclasses. Return a list of class "fa".
}
\section{Methods}{
\describe{

\item{\code{signature(obj = "Fa")}}{
Accessor method to some slots of an object of class \code{"Fa"} and its subclasses. Return a list of class "fa".
}
}}
\keyword{methods}