\name{getLoadings-methods}
\docType{methods}
\alias{getLoadings-methods}
\alias{getLoadings,Fa-method}
\title{Access Loadings slot}
\description{
Accessor method to the Loadings slot of an object of class \code{"Fa"} and its subclasses.
}
\section{Methods}{
\describe{

\item{\code{signature(obj = "Fa")}}{
Accessor method to the Loadings slot of an object of class \code{"Fa"} and its subclasses
}
}}
\keyword{methods}
