\name{getQuan-methods}
\docType{methods}
\alias{getQuan-methods}
\alias{getQuan,Fa-method}
\title{Access n.obs slot}
\description{
Accessor method to the n.obs slot of an object of class \code{"Fa"} and its subclasses.
}
\section{Methods}{
\describe{

\item{\code{signature(obj = "Fa")}}{
Accessor method to the n.obs slot of an object of class \code{"Fa"} and its subclasses
}
}}
\keyword{methods}
