\name{getScores-methods}
\docType{methods}
\alias{getScores-methods}
\alias{getScores,Fa-method}
\title{Access Scores slot}
\description{
Accessor method to the Scores slot of an object of class \code{"Fa"} and its subclasses.
}
\section{Methods}{
\describe{

\item{\code{signature(obj = "Fa")}}{
Accessor method to the Scores slot of an object of class \code{"Fa"} and its subclasses
}
}}
\keyword{methods}
