\name{getSdev-methods}
\docType{methods}
\alias{getSdev-methods}
\alias{getSdev,Fa-method}
\title{Access Standard Deviation slot}
\description{
Accessor method to the Standard Deviation slot of an object of class \code{"Fa"} and its subclasses.
}
\section{Methods}{
\describe{

\item{\code{signature(obj = "Fa")}}{
Accessor method to the Standard Deviation slot of an object of class \code{"Fa"} and its subclasses
}
}}
\keyword{methods}
