\name{Unumeric-class}
\Rdversion{1.1}
\docType{class}
\alias{Unumeric-class}

\title{Class \code{"Unumeric"}}
\description{
Define class unions for optional slots, e.g., for definition of slots which will be computed on demand.
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "Unumeric" in the signature.
}

\references{
Zhang, Y. Y. (2013), An Object Oriented Solution for Robust Factor Analysis.
}

\author{ Ying-Ying Zhang (Robert) \email{robertzhangyying@qq.com} }

\keyword{classes}