\name{robustfa-package}
\alias{robustfa-package}
\alias{robustfa}
\docType{package}
\title{
An Object Oriented Framework for Robust Factor Analysis
}
\description{
An object oriented framework for robust factor analysis. In the framework, new S4 classes \code{"Fa"}, \code{"FaClassic"}, \code{"FaRobust"}, \code{"FaCov"}, \code{"SummaryFa"} are created.
}
\details{
\tabular{ll}{
Package: \tab robustfa\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-02-15\cr
License: \tab GPL (>= 2)\cr
Depends: \tab methods\cr
}
The most important functions are:
\code{\link{FaClassic}}, \code{\link{FaCov}}, \code{\link{factorScorePca}}, \code{\link{factorScorePfa}}
}
\author{
Ying-Ying Zhang (Robert)

Maintainer: Ying-Ying Zhang (Robert) <robertzhang@cqu.edu.cn>
}
\references{
Zhang, Y. Y. (2012), An Object Oriented Framework for Robust Factor Analysis.
}
\keyword{ package }
%% \seealso{}
\examples{
library("robustfa")
## see the relevant functions in \code{Details}
}
